aws/bedrock-agents-poc/requirements.txt
aws-cdk-lib
aws-glue-libs
cdk_nag
constructs>=10.0.0,<11.0.0
pyspark
